requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
mariadb==1.1.8
mysql-connector==2.2.9
packaging==23.2
pip==23.3.1
requests==2.31.0
setuptools==57.4.0
soupsieve==2.5
urllib3==2.0.7